(very) rough initial draft of UserToken class
I mean very rough. My code writing is not the best, so I hope I can be forgiven... but hopefully this is a decent framework. 

One thing I'm not sure about is how we're going to deal with passing information into these methods. The interface doesn't user arguments... so I'm assuming we can't use them for the class. 
If this makes more sense, what I'd like to know is, how can we make it so that the username in the method is recognized as the username being passed in?

Again, I apologize for the caveman code. Multiple dependent files make my head hurt. I'll try to keep looking at it, myself.